feat: Re-implement set filters count
