+++ packages/paperless_api/lib/src/models/models.dart
@@ -1,32 +1,33 @@
+export 'bulk_edit_model.dart';
+export 'document_filter.dart';
+export 'document_meta_data_model.dart';
+export 'document_model.dart';
+export 'field_suggestions.dart';
+export 'filter_rule_model.dart';
+export 'group_model.dart';
 export 'labels/correspondent_model.dart';
 export 'labels/document_type_model.dart';
 export 'labels/label_model.dart';
 export 'labels/matching_algorithm.dart';
 export 'labels/storage_path_model.dart';
 export 'labels/tag_model.dart';
-export 'query_parameters/id_query_parameter.dart';
-export 'query_parameters/query_type.dart';
-export 'query_parameters/sort_field.dart';
-export 'query_parameters/sort_order.dart';
-export 'query_parameters/date_range_queries/date_range_queries.dart';
-export 'query_parameters/tags_query/tags_query.dart';
-export 'query_parameters/text_query.dart';
-export 'bulk_edit_model.dart';
-export 'document_filter.dart';
-export 'document_meta_data_model.dart';
-export 'document_model.dart';
-export 'filter_rule_model.dart';
 export 'paged_search_result.dart';
 export 'paperless_server_exception.dart';
 export 'paperless_server_information_model.dart';
 export 'paperless_server_statistics_model.dart';
+export 'permissions/inherited_permissions.dart';
+export 'permissions/permissions.dart';
+export 'permissions/user_permission_extension.dart';
+export 'permissions/user_permissions.dart';
+export 'query_parameters/date_range_queries/date_range_query.dart';
+export 'query_parameters/date_range_queries/date_range_unit.dart';
+export 'query_parameters/id_query_parameter.dart';
+export 'query_parameters/query_type.dart';
+export 'query_parameters/sort_field.dart';
+export 'query_parameters/sort_order.dart';
+export 'query_parameters/tags_query/tags_query.dart';
+export 'query_parameters/text_query.dart';
 export 'saved_view_model.dart';
 export 'task/task.dart';
 export 'task/task_status.dart';
-export 'field_suggestions.dart';
-export 'permissions/permissions.dart';
-export 'permissions/user_permissions.dart';
-export 'permissions/inherited_permissions.dart';
-export 'group_model.dart';
 export 'user_model.dart';
-export 'permissions/user_permission_extension.dart';
