+++ packages/paperless_api/pubspec.yaml
@@ -23,9 +23,6 @@ dependencies:
   freezed_annotation: ^2.2.0
   hive: ^2.2.3
 
-dependency_overrides:
-  graphs: 2.2.0
-
 dev_dependencies:
   flutter_test:
     sdk: flutter
