+++ pubspec.yaml
@@ -93,7 +93,6 @@ dependencies:
 
 dependency_overrides:
   intl: ^0.18.0
-  graphs: 2.2.0 
   http: ^1.0.0
